Basic Flow for the project completed

+++ OnlineStore/store/models.py
@@ -4,5 +4,8 @@ from mongoengine import *
 class Store(Document):
     name = StringField(max_length=200)
     address = StringField(max_length=500)
-    latitude = DecimalField(max_digits=9, decimal_places=6)
-    longitude = DecimalField(max_digits=9, decimal_places=6)
+    latitude = FloatField(max_digits=15, decimal_places=10)
+    longitude = FloatField(max_digits=15, decimal_places=10)
+
+
+
